trace_limit을 환경변수로 설정하도록 하고 기본값 끔
